requirements.txt
flask
feedparser
pyyaml
apscheduler
requests
